The Project
This is a copy of the project that my team and I made for out first Hackathon, HowdyHack 2020. It was a ton of work -  spent around 18 of the 24 hours at the computer - but I learned a ton! To see the original project, go to https://github.com/robmadriaga/Extra_Secret